pyproject.toml
[project]
name = "cv-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.5.6",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "huggingface-hub>=1.5.0",
    "openai>=2.24.0",
    "passlib[bcrypt]>=1.7.4",
    "pbkdf2>=1.3",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.54.0",
    "tavily-python>=0.7.22",
    "uvicorn>=0.41.0",
]
